+++ lib/elixir_script/prelude/kernel.ex
@@ -1,6 +1,7 @@
 defmodule ElixirScript.Kernel do
   import Kernel, only: [defmodule: 2, def: 1, def: 2, defp: 2,
-  defmacro: 1, defmacro: 2, defmacrop: 2, ||: 2, !: 1, ++: 2, in: 2, &&: 2, ===: 2, @: 1]
+                        defmacro: 1, defmacro: 2, defmacrop: 2, ||: 2, !: 1, ++: 2, in: 2, &&: 2, ===: 2, @: 1]
+  require JS
 
 
   defmacro if(condition, clauses) do